% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{add_role_to_db_instance}
\alias{add_role_to_db_instance}
\title{Associates an AWS Identity and Access Management (IAM) role with a DB instance}
\usage{
add_role_to_db_instance(DBInstanceIdentifier, RoleArn, FeatureName)
}
\arguments{
\item{DBInstanceIdentifier}{[required] The name of the DB instance to associate the IAM role with.}

\item{RoleArn}{[required] The Amazon Resource Name (ARN) of the IAM role to associate with the DB instance, for example \code{arn:aws:iam::123456789012:role/AccessRole}.}

\item{FeatureName}{[required] The name of the feature for the DB instance that the IAM role is to be associated with. For the list of supported feature names, see DBEngineVersion.}
}
\description{
Associates an AWS Identity and Access Management (IAM) role with a DB instance.
}
\section{Accepted Parameters}{
\preformatted{add_role_to_db_instance(
  DBInstanceIdentifier = "string",
  RoleArn = "string",
  FeatureName = "string"
)
}
}

