% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{add_source_identifier_to_subscription}
\alias{add_source_identifier_to_subscription}
\title{Adds a source identifier to an existing RDS event notification subscription}
\usage{
add_source_identifier_to_subscription(
  SubscriptionName = "string",
  SourceIdentifier = "string"
)
}
\arguments{
\item{SubscriptionName}{[required] The name of the RDS event notification subscription you want to add a source identifier to.}

\item{SourceIdentifier}{[required] The identifier of the event source to be added.

Constraints:
\itemize{
\item If the source type is a DB instance, then a \code{DBInstanceIdentifier} must be supplied.
\item If the source type is a DB security group, a \code{DBSecurityGroupName} must be supplied.
\item If the source type is a DB parameter group, a \code{DBParameterGroupName} must be supplied.
\item If the source type is a DB snapshot, a \code{DBSnapshotIdentifier} must be supplied.
}}
}
\description{
Adds a source identifier to an existing RDS event notification subscription.
}
\examples{

}
