% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{authorize_db_security_group_ingress}
\alias{authorize_db_security_group_ingress}
\title{Enables ingress to a DBSecurityGroup using one of two forms of authorization}
\usage{
authorize_db_security_group_ingress(DBSecurityGroupName, CIDRIP = NULL,
  EC2SecurityGroupName = NULL, EC2SecurityGroupId = NULL,
  EC2SecurityGroupOwnerId = NULL)
}
\arguments{
\item{DBSecurityGroupName}{The name of the DB security group to add authorization to.}

\item{CIDRIP}{The IP range to authorize.}

\item{EC2SecurityGroupName}{Name of the EC2 security group to authorize. For VPC DB security groups, \code{EC2SecurityGroupId} must be provided. Otherwise, \code{EC2SecurityGroupOwnerId} and either \code{EC2SecurityGroupName} or \code{EC2SecurityGroupId} must be provided.}

\item{EC2SecurityGroupId}{Id of the EC2 security group to authorize. For VPC DB security groups, \code{EC2SecurityGroupId} must be provided. Otherwise, \code{EC2SecurityGroupOwnerId} and either \code{EC2SecurityGroupName} or \code{EC2SecurityGroupId} must be provided.}

\item{EC2SecurityGroupOwnerId}{AWS account number of the owner of the EC2 security group specified in the \code{EC2SecurityGroupName} parameter. The AWS Access Key ID is not an acceptable value. For VPC DB security groups, \code{EC2SecurityGroupId} must be provided. Otherwise, \code{EC2SecurityGroupOwnerId} and either \code{EC2SecurityGroupName} or \code{EC2SecurityGroupId} must be provided.}
}
\description{
Enables ingress to a DBSecurityGroup using one of two forms of authorization. First, EC2 or VPC security groups can be added to the DBSecurityGroup if the application using the database is running on EC2 or VPC instances. Second, IP ranges are available if the application accessing your database is running on the Internet. Required parameters for this API are one of CIDR range, EC2SecurityGroupId for VPC, or (EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId for non-VPC).
}
\details{
You can't authorize ingress from an EC2 security group in one AWS Region to an Amazon RDS DB instance in another. You can't authorize ingress from a VPC security group in one VPC to an Amazon RDS DB instance in another.

For an overview of CIDR ranges, go to the \href{http://en.wikipedia.org/wiki/Classless_Inter-Domain_Routing}{Wikipedia Tutorial}.
}
\examples{

}
