% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{create_db_security_group}
\alias{create_db_security_group}
\title{Creates a new DB security group}
\usage{
create_db_security_group(
  DBSecurityGroupName = "string",
  DBSecurityGroupDescription = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
\arguments{
\item{DBSecurityGroupName}{[required] The name for the DB security group. This value is stored as a lowercase string.

Constraints:
\itemize{
\item Must be 1 to 255 letters, numbers, or hyphens.
\item First character must be a letter
\item Can't end with a hyphen or contain two consecutive hyphens
\item Must not be "Default"
}

Example: \code{mysecuritygroup}}

\item{DBSecurityGroupDescription}{[required] The description for the DB security group.}

\item{Tags}{}
}
\description{
Creates a new DB security group. DB security groups control access to a DB instance.
}
\details{
A DB security group controls access to EC2-Classic DB instances that are not in a VPC.
}
\examples{

}
