% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{create_db_snapshot}
\alias{create_db_snapshot}
\title{Creates a DBSnapshot}
\usage{
create_db_snapshot(DBSnapshotIdentifier, DBInstanceIdentifier,
  Tags = NULL)
}
\arguments{
\item{DBSnapshotIdentifier}{[required] The identifier for the DB snapshot.

Constraints:
\itemize{
\item Can't be null, empty, or blank
\item Must contain from 1 to 255 letters, numbers, or hyphens
\item First character must be a letter
\item Can't end with a hyphen or contain two consecutive hyphens
}

Example: \code{my-snapshot-id}}

\item{DBInstanceIdentifier}{[required] The identifier of the DB instance that you want to create the snapshot of.

Constraints:
\itemize{
\item Must match the identifier of an existing DBInstance.
}}

\item{Tags}{}
}
\description{
Creates a DBSnapshot. The source DBInstance must be in "available" state.
}
\section{Accepted Parameters}{
\preformatted{create_db_snapshot(
  DBSnapshotIdentifier = "string",
  DBInstanceIdentifier = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

