% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{create_db_subnet_group}
\alias{create_db_subnet_group}
\title{Creates a new DB subnet group}
\usage{
create_db_subnet_group(DBSubnetGroupName, DBSubnetGroupDescription, SubnetIds,
  Tags = NULL)
}
\arguments{
\item{DBSubnetGroupName}{The name for the DB subnet group. This value is stored as a lowercase string.

Constraints: Must contain no more than 255 letters, numbers, periods, underscores, spaces, or hyphens. Must not be default.

Example: \code{mySubnetgroup}}

\item{DBSubnetGroupDescription}{The description for the DB subnet group.}

\item{SubnetIds}{The EC2 Subnet IDs for the DB subnet group.}

\item{Tags}{}
}
\description{
Creates a new DB subnet group. DB subnet groups must contain at least one subnet in at least two AZs in the AWS Region.
}
\examples{

}
