% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{create_option_group}
\alias{create_option_group}
\title{Creates a new option group}
\usage{
create_option_group(OptionGroupName, EngineName, MajorEngineVersion,
  OptionGroupDescription, Tags = NULL)
}
\arguments{
\item{OptionGroupName}{Specifies the name of the option group to be created.

Constraints:
\itemize{
\item Must be 1 to 255 letters, numbers, or hyphens
\item First character must be a letter
\item Can't end with a hyphen or contain two consecutive hyphens
}

Example: \code{myoptiongroup}}

\item{EngineName}{Specifies the name of the engine that this option group should be associated with.}

\item{MajorEngineVersion}{Specifies the major version of the engine that this option group should be associated with.}

\item{OptionGroupDescription}{The description of the option group.}

\item{Tags}{}
}
\description{
Creates a new option group. You can create up to 20 option groups.
}
\examples{

}
