% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{delete_db_cluster_parameter_group}
\alias{delete_db_cluster_parameter_group}
\title{Deletes a specified DB cluster parameter group}
\usage{
delete_db_cluster_parameter_group(
  DBClusterParameterGroupName = "string"
)
}
\arguments{
\item{DBClusterParameterGroupName}{[required] The name of the DB cluster parameter group.

Constraints:
\itemize{
\item Must be the name of an existing DB cluster parameter group.
\item You can't delete a default DB cluster parameter group.
\item Can't be associated with any DB clusters.
}}
}
\description{
Deletes a specified DB cluster parameter group. The DB cluster parameter group to be deleted can't be associated with any DB clusters.
}
\details{
For more information on Amazon Aurora, see \href{http://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/CHAP_AuroraOverview.html}{What Is Amazon Aurora?} in the \emph{Amazon Aurora User Guide.}
}
\examples{

}
