% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{delete_db_cluster_snapshot}
\alias{delete_db_cluster_snapshot}
\title{Deletes a DB cluster snapshot}
\usage{
delete_db_cluster_snapshot(DBClusterSnapshotIdentifier)
}
\arguments{
\item{DBClusterSnapshotIdentifier}{The identifier of the DB cluster snapshot to delete.

Constraints: Must be the name of an existing DB cluster snapshot in the \code{available} state.}
}
\description{
Deletes a DB cluster snapshot. If the snapshot is being copied, the copy operation is terminated.
}
\details{
The DB cluster snapshot must be in the \code{available} state to be deleted.

For more information on Amazon Aurora, see \href{http://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/CHAP_AuroraOverview.html}{What Is Amazon Aurora?} in the \emph{Amazon Aurora User Guide.}
}
\examples{

}
