% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{delete_db_instance_automated_backup}
\alias{delete_db_instance_automated_backup}
\title{Deletes automated backups based on the source instance's \code{DbiResourceId} value or the restorable instance's resource ID}
\usage{
delete_db_instance_automated_backup(
  DbiResourceId = "string"
)
}
\arguments{
\item{DbiResourceId}{[required] The identifier for the source DB instance, which can't be changed and which is unique to an AWS Region.}
}
\description{
Deletes automated backups based on the source instance's \code{DbiResourceId} value or the restorable instance's resource ID.
}
\examples{

}
