% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{delete_db_parameter_group}
\alias{delete_db_parameter_group}
\title{Deletes a specified DB parameter group}
\usage{
delete_db_parameter_group(DBParameterGroupName)
}
\arguments{
\item{DBParameterGroupName}{[required] The name of the DB parameter group.

Constraints:
\itemize{
\item Must be the name of an existing DB parameter group
\item You can't delete a default DB parameter group
\item Can't be associated with any DB instances
}}
}
\description{
Deletes a specified DB parameter group. The DB parameter group to be deleted can't be associated with any DB instances.
}
\section{Accepted Parameters}{
\preformatted{delete_db_parameter_group(
  DBParameterGroupName = "string"
)
}
}

