% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{delete_db_snapshot}
\alias{delete_db_snapshot}
\title{Deletes a DB snapshot}
\usage{
delete_db_snapshot(DBSnapshotIdentifier)
}
\arguments{
\item{DBSnapshotIdentifier}{[required] The DB snapshot identifier.

Constraints: Must be the name of an existing DB snapshot in the \code{available} state.}
}
\description{
Deletes a DB snapshot. If the snapshot is being copied, the copy operation is terminated.
}
\details{
The DB snapshot must be in the \code{available} state to be deleted.
}
\section{Accepted Parameters}{
\preformatted{delete_db_snapshot(
  DBSnapshotIdentifier = "string"
)
}
}

