% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{delete_event_subscription}
\alias{delete_event_subscription}
\title{Deletes an RDS event notification subscription}
\usage{
delete_event_subscription(
  SubscriptionName = "string"
)
}
\arguments{
\item{SubscriptionName}{[required] The name of the RDS event notification subscription you want to delete.}
}
\description{
Deletes an RDS event notification subscription.
}
\examples{

}
