% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{delete_global_cluster}
\alias{delete_global_cluster}
\title{Deletes a global database cluster}
\usage{
delete_global_cluster(GlobalClusterIdentifier)
}
\arguments{
\item{GlobalClusterIdentifier}{[required] The cluster identifier of the global database cluster being deleted.}
}
\description{
Deletes a global database cluster. The primary and secondary clusters must already be detached or destroyed first.
}
\section{Accepted Parameters}{
\preformatted{delete_global_cluster(
  GlobalClusterIdentifier = "string"
)
}
}

