% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{delete_option_group}
\alias{delete_option_group}
\title{Deletes an existing option group}
\usage{
delete_option_group(OptionGroupName)
}
\arguments{
\item{OptionGroupName}{The name of the option group to be deleted.

You can't delete default option groups.}
}
\description{
Deletes an existing option group.
}
\examples{

}
