% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{describe_account_attributes}
\alias{describe_account_attributes}
\title{Lists all of the attributes for a customer account}
\usage{
describe_account_attributes()
}
\description{
Lists all of the attributes for a customer account. The attributes include Amazon RDS quotas for the account, such as the number of DB instances allowed. The description for a quota includes the quota name, current usage toward that quota, and the quota's maximum value.
}
\details{
This command doesn't take any parameters.
}
\examples{

}
