% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{describe_event_categories}
\alias{describe_event_categories}
\title{Displays a list of categories for all event source types, or, if specified, for a specified source type}
\usage{
describe_event_categories(SourceType = NULL, Filters = NULL)
}
\arguments{
\item{SourceType}{The type of source that is generating the events.

Valid values: db-instance | db-parameter-group | db-security-group | db-snapshot}

\item{Filters}{This parameter is not currently supported.}
}
\description{
Displays a list of categories for all event source types, or, if specified, for a specified source type. You can see a list of the event categories and source types in the \href{http://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_Events.html}{Events} topic in the \emph{Amazon RDS User Guide.}
}
\section{Accepted Parameters}{
\preformatted{describe_event_categories(
  SourceType = "string",
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  )
)
}
}

