% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{describe_option_groups}
\alias{describe_option_groups}
\title{Describes the available option groups}
\usage{
describe_option_groups(OptionGroupName = NULL, Filters = NULL,
  Marker = NULL, MaxRecords = NULL, EngineName = NULL,
  MajorEngineVersion = NULL)
}
\arguments{
\item{OptionGroupName}{The name of the option group to describe. Can't be supplied together with EngineName or MajorEngineVersion.}

\item{Filters}{This parameter is not currently supported.}

\item{Marker}{An optional pagination token provided by a previous DescribeOptionGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by \code{MaxRecords}.}

\item{MaxRecords}{The maximum number of records to include in the response. If more records exist than the specified \code{MaxRecords} value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{EngineName}{Filters the list of option groups to only include groups associated with a specific database engine.}

\item{MajorEngineVersion}{Filters the list of option groups to only include groups associated with a specific database engine version. If specified, then EngineName must also be specified.}
}
\description{
Describes the available option groups.
}
\examples{

}
