% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{describe_source_regions}
\alias{describe_source_regions}
\title{Returns a list of the source AWS Regions where the current AWS Region can create a Read Replica or copy a DB snapshot from}
\usage{
describe_source_regions(RegionName = NULL, MaxRecords = NULL,
  Marker = NULL, Filters = NULL)
}
\arguments{
\item{RegionName}{The source AWS Region name. For example, \code{us-east-1}.

Constraints:
\itemize{
\item Must specify a valid AWS Region name.
}}

\item{MaxRecords}{The maximum number of records to include in the response. If more records exist than the specified \code{MaxRecords} value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous DescribeSourceRegions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by \code{MaxRecords}.}

\item{Filters}{This parameter is not currently supported.}
}
\description{
Returns a list of the source AWS Regions where the current AWS Region can create a Read Replica or copy a DB snapshot from. This API action supports pagination.
}
\section{Accepted Parameters}{
\preformatted{describe_source_regions(
  RegionName = "string",
  MaxRecords = 123,
  Marker = "string",
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  )
)
}
}

