% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{describe_valid_db_instance_modifications}
\alias{describe_valid_db_instance_modifications}
\title{You can call DescribeValidDBInstanceModifications to learn what modifications you can make to your DB instance}
\usage{
describe_valid_db_instance_modifications(
  DBInstanceIdentifier = "string"
)
}
\arguments{
\item{DBInstanceIdentifier}{[required] The customer identifier or the ARN of your DB instance.}
}
\description{
You can call DescribeValidDBInstanceModifications to learn what modifications you can make to your DB instance. You can use this information when you call ModifyDBInstance.
}
\examples{

}
