% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{modify_db_cluster_snapshot_attribute}
\alias{modify_db_cluster_snapshot_attribute}
\title{Adds an attribute and values to, or removes an attribute and values from, a manual DB cluster snapshot}
\usage{
modify_db_cluster_snapshot_attribute(DBClusterSnapshotIdentifier, AttributeName,
  ValuesToAdd = NULL, ValuesToRemove = NULL)
}
\arguments{
\item{DBClusterSnapshotIdentifier}{The identifier for the DB cluster snapshot to modify the attributes for.}

\item{AttributeName}{The name of the DB cluster snapshot attribute to modify.

To manage authorization for other AWS accounts to copy or restore a manual DB cluster snapshot, set this value to \code{restore}.}

\item{ValuesToAdd}{A list of DB cluster snapshot attributes to add to the attribute specified by \code{AttributeName}.

To authorize other AWS accounts to copy or restore a manual DB cluster snapshot, set this list to include one or more AWS account IDs, or \code{all} to make the manual DB cluster snapshot restorable by any AWS account. Do not add the \code{all} value for any manual DB cluster snapshots that contain private information that you don't want available to all AWS accounts.}

\item{ValuesToRemove}{A list of DB cluster snapshot attributes to remove from the attribute specified by \code{AttributeName}.

To remove authorization for other AWS accounts to copy or restore a manual DB cluster snapshot, set this list to include one or more AWS account identifiers, or \code{all} to remove authorization for any AWS account to copy or restore the DB cluster snapshot. If you specify \code{all}, an AWS account whose account ID is explicitly added to the \code{restore} attribute can still copy or restore a manual DB cluster snapshot.}
}
\description{
Adds an attribute and values to, or removes an attribute and values from, a manual DB cluster snapshot.
}
\details{
To share a manual DB cluster snapshot with other AWS accounts, specify \code{restore} as the \code{AttributeName} and use the \code{ValuesToAdd} parameter to add a list of IDs of the AWS accounts that are authorized to restore the manual DB cluster snapshot. Use the value \code{all} to make the manual DB cluster snapshot public, which means that it can be copied or restored by all AWS accounts. Do not add the \code{all} value for any manual DB cluster snapshots that contain private information that you don't want available to all AWS accounts. If a manual DB cluster snapshot is encrypted, it can be shared, but only by specifying a list of authorized AWS account IDs for the \code{ValuesToAdd} parameter. You can't use \code{all} as a value for that parameter in this case.

To view which AWS accounts have access to copy or restore a manual DB cluster snapshot, or whether a manual DB cluster snapshot public or private, use the DescribeDBClusterSnapshotAttributes API action.
}
\examples{

}
