% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{modify_db_subnet_group}
\alias{modify_db_subnet_group}
\title{Modifies an existing DB subnet group}
\usage{
modify_db_subnet_group(DBSubnetGroupName, SubnetIds,
  DBSubnetGroupDescription = NULL)
}
\arguments{
\item{DBSubnetGroupName}{The name for the DB subnet group. This value is stored as a lowercase string. You can't modify the default subnet group.

Constraints: Must match the name of an existing DBSubnetGroup. Must not be default.

Example: \code{mySubnetgroup}}

\item{SubnetIds}{The EC2 subnet IDs for the DB subnet group.}

\item{DBSubnetGroupDescription}{The description for the DB subnet group.}
}
\description{
Modifies an existing DB subnet group. DB subnet groups must contain at least one subnet in at least two AZs in the AWS Region.
}
\examples{

}
