% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{modify_option_group}
\alias{modify_option_group}
\title{Modifies an existing option group}
\usage{
modify_option_group(OptionGroupName, OptionsToInclude = NULL,
  OptionsToRemove = NULL, ApplyImmediately = NULL)
}
\arguments{
\item{OptionGroupName}{[required] The name of the option group to be modified.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, can't be removed from an option group, and that option group can't be removed from a DB instance once it is associated with a DB instance}

\item{OptionsToInclude}{Options in this list are added to the option group or, if already present, the specified configuration is used to update the existing configuration.}

\item{OptionsToRemove}{Options in this list are removed from the option group.}

\item{ApplyImmediately}{Indicates whether the changes should be applied immediately, or during the next maintenance window for each instance associated with the option group.}
}
\description{
Modifies an existing option group.
}
\section{Accepted Parameters}{
\preformatted{modify_option_group(
  OptionGroupName = "string",
  OptionsToInclude = list(
    list(
      OptionName = "string",
      Port = 123,
      OptionVersion = "string",
      DBSecurityGroupMemberships = list(
        "string"
      ),
      VpcSecurityGroupMemberships = list(
        "string"
      ),
      OptionSettings = list(
        list(
          Name = "string",
          Value = "string",
          DefaultValue = "string",
          Description = "string",
          ApplyType = "string",
          DataType = "string",
          AllowedValues = "string",
          IsModifiable = TRUE|FALSE,
          IsCollection = TRUE|FALSE
        )
      )
    )
  ),
  OptionsToRemove = list(
    "string"
  ),
  ApplyImmediately = TRUE|FALSE
)
}
}

