% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_package.R
\docType{package}
\name{paws.rds-package}
\alias{paws.rds}
\alias{paws.rds-package}
\title{paws.rds: Amazon Relational Database Service}
\description{
Amazon Relational Database Service (Amazon RDS) is
a web service that makes it easier to set up, operate, and scale a
relational database in the cloud. It provides cost-efficient,
resizable capacity for an industry-standard relational database and
manages common database administration tasks, freeing up developers to
focus on what makes their applications and businesses unique.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
