% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{promote_read_replica_db_cluster}
\alias{promote_read_replica_db_cluster}
\title{Promotes a Read Replica DB cluster to a standalone DB cluster}
\usage{
promote_read_replica_db_cluster(DBClusterIdentifier)
}
\arguments{
\item{DBClusterIdentifier}{The identifier of the DB cluster Read Replica to promote. This parameter is not case-sensitive.

Constraints:
\itemize{
\item Must match the identifier of an existing DBCluster Read Replica.
}

Example: \code{my-cluster-replica1}}
}
\description{
Promotes a Read Replica DB cluster to a standalone DB cluster.
}
\examples{

}
