% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{purchase_reserved_db_instances_offering}
\alias{purchase_reserved_db_instances_offering}
\title{Purchases a reserved DB instance offering}
\usage{
purchase_reserved_db_instances_offering(
  ReservedDBInstancesOfferingId = "string",
  ReservedDBInstanceId = "string",
  DBInstanceCount = 123,
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
\arguments{
\item{ReservedDBInstancesOfferingId}{[required] The ID of the Reserved DB instance offering to purchase.

Example: 438012d3-4052-4cc7-b2e3-8d3372e0e706}

\item{ReservedDBInstanceId}{Customer-specified identifier to track this reservation.

Example: myreservationID}

\item{DBInstanceCount}{The number of instances to reserve.

Default: \code{1}}

\item{Tags}{}
}
\description{
Purchases a reserved DB instance offering.
}
\examples{

}
