% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{remove_from_global_cluster}
\alias{remove_from_global_cluster}
\title{Detaches an Aurora secondary cluster from an Aurora global database cluster}
\usage{
remove_from_global_cluster(GlobalClusterIdentifier = NULL,
  DbClusterIdentifier = NULL)
}
\arguments{
\item{GlobalClusterIdentifier}{The cluster identifier to detach from the Aurora global database cluster.}

\item{DbClusterIdentifier}{The Amazon Resource Name (ARN) identifying the cluster that was detached from the Aurora global database cluster.}
}
\description{
Detaches an Aurora secondary cluster from an Aurora global database cluster. The cluster becomes a standalone cluster with read-write capability instead of being read-only and receiving data from a primary cluster in a different region.
}
\section{Accepted Parameters}{
\preformatted{remove_from_global_cluster(
  GlobalClusterIdentifier = "string",
  DbClusterIdentifier = "string"
)
}
}

