% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{remove_role_from_db_cluster}
\alias{remove_role_from_db_cluster}
\title{Disassociates an Identity and Access Management (IAM) role from an Aurora DB cluster}
\usage{
remove_role_from_db_cluster(DBClusterIdentifier, RoleArn)
}
\arguments{
\item{DBClusterIdentifier}{The name of the DB cluster to disassociate the IAM role from.}

\item{RoleArn}{The Amazon Resource Name (ARN) of the IAM role to disassociate from the Aurora DB cluster, for example \code{arn:aws:iam::123456789012:role/AuroraAccessRole}.}
}
\description{
Disassociates an Identity and Access Management (IAM) role from an Aurora DB cluster. For more information, see \href{http://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/AuroraMySQL.Integrating.Authorizing.html}{Authorizing Amazon Aurora MySQL to Access Other AWS Services on Your Behalf} in the \emph{Amazon Aurora User Guide}.
}
\examples{

}
