% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{reset_db_parameter_group}
\alias{reset_db_parameter_group}
\title{Modifies the parameters of a DB parameter group to the engine/system default value}
\usage{
reset_db_parameter_group(DBParameterGroupName, ResetAllParameters = NULL,
  Parameters = NULL)
}
\arguments{
\item{DBParameterGroupName}{The name of the DB parameter group.

Constraints:
\itemize{
\item Must match the name of an existing DBParameterGroup.
}}

\item{ResetAllParameters}{Specifies whether (\code{true}) or not (\code{false}) to reset all parameters in the DB parameter group to default values.

Default: \code{true}}

\item{Parameters}{To reset the entire DB parameter group, specify the \code{DBParameterGroup} name and \code{ResetAllParameters} parameters. To reset specific parameters, provide a list of the following: \code{ParameterName} and \code{ApplyMethod}. A maximum of 20 parameters can be modified in a single request.

\strong{MySQL}

Valid Values (for Apply method): \code{immediate} | \code{pending-reboot}

You can use the immediate value with dynamic parameters only. You can use the \code{pending-reboot} value for both dynamic and static parameters, and changes are applied when DB instance reboots.

\strong{MariaDB}

Valid Values (for Apply method): \code{immediate} | \code{pending-reboot}

You can use the immediate value with dynamic parameters only. You can use the \code{pending-reboot} value for both dynamic and static parameters, and changes are applied when DB instance reboots.

\strong{Oracle}

Valid Values (for Apply method): \code{pending-reboot}}
}
\description{
Modifies the parameters of a DB parameter group to the engine/system default value. To reset specific parameters, provide a list of the following: \code{ParameterName} and \code{ApplyMethod}. To reset the entire DB parameter group, specify the \code{DBParameterGroup} name and \code{ResetAllParameters} parameters. When resetting the entire group, dynamic parameters are updated immediately and static parameters are set to \code{pending-reboot} to take effect on the next DB instance restart or \code{RebootDBInstance} request.
}
\examples{

}
