% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{restore_db_cluster_from_s3}
\alias{restore_db_cluster_from_s3}
\title{Creates an Amazon Aurora DB cluster from data stored in an Amazon S3 bucket}
\usage{
restore_db_cluster_from_s3(
  AvailabilityZones = list(
    "string"
  ),
  BackupRetentionPeriod = 123,
  CharacterSetName = "string",
  DatabaseName = "string",
  DBClusterIdentifier = "string",
  DBClusterParameterGroupName = "string",
  VpcSecurityGroupIds = list(
    "string"
  ),
  DBSubnetGroupName = "string",
  Engine = "string",
  EngineVersion = "string",
  Port = 123,
  MasterUsername = "string",
  MasterUserPassword = "string",
  OptionGroupName = "string",
  PreferredBackupWindow = "string",
  PreferredMaintenanceWindow = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  StorageEncrypted = TRUE|FALSE,
  KmsKeyId = "string",
  EnableIAMDatabaseAuthentication = TRUE|FALSE,
  SourceEngine = "string",
  SourceEngineVersion = "string",
  S3BucketName = "string",
  S3Prefix = "string",
  S3IngestionRoleArn = "string",
  BacktrackWindow = 123,
  EnableCloudwatchLogsExports = list(
    "string"
  ),
  DeletionProtection = TRUE|FALSE
)
}
\arguments{
\item{AvailabilityZones}{A list of EC2 Availability Zones that instances in the restored DB cluster can be created in.}

\item{BackupRetentionPeriod}{The number of days for which automated backups of the restored DB cluster are retained. You must specify a minimum value of 1.

Default: 1

Constraints:
\itemize{
\item Must be a value from 1 to 35
}}

\item{CharacterSetName}{A value that indicates that the restored DB cluster should be associated with the specified CharacterSet.}

\item{DatabaseName}{The database name for the restored DB cluster.}

\item{DBClusterIdentifier}{[required] The name of the DB cluster to create from the source data in the Amazon S3 bucket. This parameter is isn't case-sensitive.

Constraints:
\itemize{
\item Must contain from 1 to 63 letters, numbers, or hyphens.
\item First character must be a letter.
\item Can't end with a hyphen or contain two consecutive hyphens.
}

Example: \code{my-cluster1}}

\item{DBClusterParameterGroupName}{The name of the DB cluster parameter group to associate with the restored DB cluster. If this argument is omitted, \code{default.aurora5.6} is used.

Constraints:
\itemize{
\item If supplied, must match the name of an existing DBClusterParameterGroup.
}}

\item{VpcSecurityGroupIds}{A list of EC2 VPC security groups to associate with the restored DB cluster.}

\item{DBSubnetGroupName}{A DB subnet group to associate with the restored DB cluster.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: \code{mySubnetgroup}}

\item{Engine}{[required] The name of the database engine to be used for the restored DB cluster.

Valid Values: \code{aurora}, \code{aurora-postgresql}}

\item{EngineVersion}{The version number of the database engine to use.

\strong{Aurora MySQL}

Example: \code{5.6.10a}

\strong{Aurora PostgreSQL}

Example: \code{9.6.3}}

\item{Port}{The port number on which the instances in the restored DB cluster accept connections.

Default: \code{3306}}

\item{MasterUsername}{[required] The name of the master user for the restored DB cluster.

Constraints:
\itemize{
\item Must be 1 to 16 letters or numbers.
\item First character must be a letter.
\item Can't be a reserved word for the chosen database engine.
}}

\item{MasterUserPassword}{[required] The password for the master database user. This password can contain any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.}

\item{OptionGroupName}{A value that indicates that the restored DB cluster should be associated with the specified option group.

Permanent options can't be removed from an option group. An option group can't be removed from a DB cluster once it is associated with a DB cluster.}

\item{PreferredBackupWindow}{The daily time range during which automated backups are created if automated backups are enabled using the \code{BackupRetentionPeriod} parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each AWS Region. To see the time blocks available, see \href{http://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/USER_UpgradeDBInstance.Maintenance.html#AdjustingTheMaintenanceWindow.Aurora}{Adjusting the Preferred Maintenance Window} in the \emph{Amazon Aurora User Guide.}

Constraints:
\itemize{
\item Must be in the format \code{hh24:mi-hh24:mi}.
\item Must be in Universal Coordinated Time (UTC).
\item Must not conflict with the preferred maintenance window.
\item Must be at least 30 minutes.
}}

\item{PreferredMaintenanceWindow}{The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: \code{ddd:hh24:mi-ddd:hh24:mi}

The default is a 30-minute window selected at random from an 8-hour block of time for each AWS Region, occurring on a random day of the week. To see the time blocks available, see \href{http://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/USER_UpgradeDBInstance.Maintenance.html#AdjustingTheMaintenanceWindow.Aurora}{Adjusting the Preferred Maintenance Window} in the \emph{Amazon Aurora User Guide.}

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.}

\item{Tags}{}

\item{StorageEncrypted}{Specifies whether the restored DB cluster is encrypted.}

\item{KmsKeyId}{The AWS KMS key identifier for an encrypted DB cluster.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB cluster with the same AWS account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KM encryption key.

If the \code{StorageEncrypted} parameter is true, and you do not specify a value for the \code{KmsKeyId} parameter, then Amazon RDS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS Region.}

\item{EnableIAMDatabaseAuthentication}{True to enable mapping of AWS Identity and Access Management (IAM) accounts to database accounts, and otherwise false.

Default: \code{false}}

\item{SourceEngine}{[required] The identifier for the database engine that was backed up to create the files stored in the Amazon S3 bucket.

Valid values: \code{mysql}}

\item{SourceEngineVersion}{[required] The version of the database that the backup files were created from.

MySQL version 5.5 and 5.6 are supported.

Example: \code{5.6.22}}

\item{S3BucketName}{[required] The name of the Amazon S3 bucket that contains the data used to create the Amazon Aurora DB cluster.}

\item{S3Prefix}{The prefix for all of the file names that contain the data used to create the Amazon Aurora DB cluster. If you do not specify a \strong{SourceS3Prefix} value, then the Amazon Aurora DB cluster is created by using all of the files in the Amazon S3 bucket.}

\item{S3IngestionRoleArn}{[required] The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that authorizes Amazon RDS to access the Amazon S3 bucket on your behalf.}

\item{BacktrackWindow}{The target backtrack window, in seconds. To disable backtracking, set this value to 0.

Default: 0

Constraints:
\itemize{
\item If specified, this value must be set to a number from 0 to 259,200 (72 hours).
}}

\item{EnableCloudwatchLogsExports}{The list of logs that the restored DB cluster is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see \href{http://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/USER_LogAccess.html#USER_LogAccess.Procedural.UploadtoCloudWatch}{Publishing Database Logs to Amazon CloudWatch Logs} in the \emph{Amazon Aurora User Guide}.}

\item{DeletionProtection}{Indicates if the DB cluster should have deletion protection enabled. The database can't be deleted when this value is set to true. The default is false.}
}
\description{
Creates an Amazon Aurora DB cluster from data stored in an Amazon S3 bucket. Amazon RDS must be authorized to access the Amazon S3 bucket and the data must be created using the Percona XtraBackup utility as described in \href{http://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/AuroraMySQL.Migrating.html}{Migrating Data to an Amazon Aurora MySQL DB Cluster} in the \emph{Amazon Aurora User Guide}.
}
\examples{

}
