% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{start_db_instance}
\alias{start_db_instance}
\title{Starts an Amazon RDS DB instance that was stopped using the AWS console, the stop-db-instance AWS CLI command, or the StopDBInstance action}
\usage{
start_db_instance(
  DBInstanceIdentifier = "string"
)
}
\arguments{
\item{DBInstanceIdentifier}{[required] The user-supplied instance identifier.}
}
\description{
Starts an Amazon RDS DB instance that was stopped using the AWS console, the stop-db-instance AWS CLI command, or the StopDBInstance action.
}
\details{
For more information, see \href{http://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_StartInstance.html}{Starting an Amazon RDS DB instance That Was Previously Stopped} in the \emph{Amazon RDS User Guide.}

This command doesn't apply to Aurora MySQL and Aurora PostgreSQL. For Aurora DB clusters, use StartDBCluster instead.
}
\examples{

}
