% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{stop_db_instance}
\alias{stop_db_instance}
\title{Stops an Amazon RDS DB instance}
\usage{
stop_db_instance(
  DBInstanceIdentifier = "string",
  DBSnapshotIdentifier = "string"
)
}
\arguments{
\item{DBInstanceIdentifier}{[required] The user-supplied instance identifier.}

\item{DBSnapshotIdentifier}{The user-supplied instance identifier of the DB Snapshot created immediately before the DB instance is stopped.}
}
\description{
Stops an Amazon RDS DB instance. When you stop a DB instance, Amazon RDS retains the DB instance's metadata, including its endpoint, DB parameter group, and option group membership. Amazon RDS also retains the transaction logs so you can do a point-in-time restore if necessary.
}
\details{
For more information, see \href{http://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_StopInstance.html}{Stopping an Amazon RDS DB Instance Temporarily} in the \emph{Amazon RDS User Guide.}

This command doesn't apply to Aurora MySQL and Aurora PostgreSQL. For Aurora clusters, use StopDBCluster instead.
}
\examples{

}
