% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.rds_operations.R
\name{delete_db_snapshot}
\alias{delete_db_snapshot}
\title{Deletes a DBSnapshot}
\usage{
delete_db_snapshot(DBSnapshotIdentifier)
}
\arguments{
\item{DBSnapshotIdentifier}{The DBSnapshot identifier.

Constraints: Must be the name of an existing DB snapshot in the \code{available} state.}
}
\description{
Deletes a DBSnapshot. If the snapshot is being copied, the copy operation is terminated.
}
\details{
The DBSnapshot must be in the \code{available} state to be deleted.
}
\examples{

}
